new flags conversion

+++ Cpp00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: isabelle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/30 19:18:10 by isabelle          #+#    #+#             */
-/*   Updated: 2022/07/05 21:24:14 by iren             ###   ########.fr       */
+/*   Updated: 2022/07/10 14:37:33 by iren             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,19 +15,19 @@
 
 int main(int ac, char **av)
 {
-	int	i;
-	int	j;
+	long unsigned int	i;
+	long unsigned int	j;
 
 	if (ac < 2)
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
 	else
 	{
 		i = 1;
-		while (i < ac)
+		while ((int)i < ac)
 		{
 			j = 0;
 			std::string res(av[i]);
-			while (j < (int)res.length())
+			while (j < res.length())
 			{
 				std::cout << (char)toupper(res[j]) << std::flush;
 				j++;
